Major refactor of pak building code
Moved the entire building code to CPakFile class as method 'BuildFromMap'. All major operations are commented now, as well as using the newly defined constants. Originally, this logic was contained within the entry point. Moving it to the class allows us to create multiple rpaks within the same process, this could easily be multi-threaded now as well. All the method takes, is a path to a valid map document.