fix page data alignment
previously any data added into the page would not always be aligned, as any data written before it may not necessarily be aligned to the same boundary as the page itself.
this is fixed by adding a fake data chunk to the page when adding a real data chunk so that the page writing function knows to skip ahead by the required number of bytes to get back to the chunk's alignment.

it is important to note that not all data within the page needs to be aligned to the whole of the page's alignment, since as long as each data chunk is independently aligned within the page, everything should work properly