shader sets
add support for shaders and shadersets on both r2 and r5

shaderset ("shds") fields:
REQUIRED:
"vertexShader" - string GUID or asset path for shader set's vertex shader
"pixelShader" - string GUID or asset path for shader set's pixel shader

OPTIONAL:
"numVertexShaderTextures" - override field for number of textures used in the vertex shader.
"numPixelShaderTextures" - override field for number of textures used in the pixel shader.
these two fields are normally automatically set by parsing the shader bytecode, but can be overridden using the provided fields.
if the vertex shader or pixel shader are external assets (i.e., not in the same pak as the shader set), these fields must be set manually, or a default value of zero is used instead, since RePak is unable to parse the bytecode of the shaders in this case
"numSamplers" - number of samplers used between the two shaders. the value of this field is not intuitive, so when porting from another version of a game, this should be taken from the original asset instead of trying to determine it yourself. in future versions, i hope to be able to automatically calculate the value of this more reliably
"firstResource" - bind point for the first (non-texture) material-controlled shader resource. this is not available in regular shaders or materials, and is only used in the UI shaders
"numResources" - number of (non-texture) material-controlled shader resources. see note above about usage.
for these resource fields, it is possible that the engine will not error when these are set incorrectly, however i wouldn't rely on it