Rework BinaryIO class
The previous implementation wasn't able to open a stream in read and write mode, as it used 2 separate streams for read and write operations. This has now been unified into a single fstream object which can do both read, write, and readwrite. This is necessary as we need read and write capabilities to perform streaming compression.

The class now also keep the size of the file without having to seek towards the end of a file to get this, seeking to end can be an expensive operation.

Also added error checking everywhere in code that uses BinaryIO to inform the user an open has failed rather than crashing the application.