Invert AddTextureAsset parameter
This should've been commited with ef4675e89ef2d1d663de856852eee53d8d2a6267 as we there explicitly error if txtr was added already since we couldn't check afterwards if its streaming unless we do a large rework there. It must be added with the uimg asset.

The naming of the parameter contradicts the operation, so it had to be flipped in the function along with all its usecases.