Fix regressions in material code
After the change forcing the entire path is to be provided (including the material/ path and .rpak extention), the partial name name was written as the full name. This has now been reworked to always write out the partial name. The CPU path was also incorrectly formatted, as that one did take the partial name but it should've been the full name minus extention.