Detect physics size mismatch
Some .phy files were exported with a truncated byte, the last section in a .phy file is the text section, and it has to be null terminated. since this byte is missing, the string is not null temrinated when loaded into the runtime. Additionally, loading it will always read 1 byte outside the phy section and when we parse the string we might parse way more than just the string due to the missing char.

Added a check to detect any size mismatches which could crash the runtime.