Implement optional streaming sets, and write all streaming sets out in parallel with the rpak
Before this patch, all data blocks were kept in memory and written to the streaming set after the pak was build, however this will result in high memory usage when building a pak with many large streaming assets. Furthermore, we had to reallocate the model buffer each time it didn't align to a 4096 bytes boundary.

We now write off the data directly into a stream which gets opened before the pak gets built, this allows us to write the data off directly while also padding the remainder of the 4096 bytes boundary out without reallocating anything.

This also allows us to immediately drop the streamed data once it has been written off.

This patch also fully implements the optional streaming set; larger textures marked optional will now be written out in the optional streaming file which allows for shipping the game without the optional textures and saving a lot of disk space and bandwidth for users that don't need their games to look as pretty as it could get.