Pak class cleanup
Attempt to open the rpak file stream before processing any assets, as otherwise we might waste time if we processed the assets and happen to find out we can't open a stream to write them out.

Also added an error if a released chunk gets written, this will probably never happen but its better to error out in cases it happens rather than a silent failure in release builds.

Also emplaced objects and modify the instance directly rather than copying the stack instance. This helps clarifying the code a bit.