Rework chunk creatin for model, animrig, animseqs
For models, the name and guid refs are always stored in 1 chunk, so we only ever create 1 CPU chunk for it now.
For animrigs, the name, rmdl data and guid refs are always stored in 1 chunk, but only the guid refs are aligned to 8 bytes. We now also only ever create 1 CPU chunk for this.
For animseqs, name and rmdl data reside in a single CPU chunk, that is aligned to 1 byte. Previously this was 64, but this is incorrect.

The resulting pak file is slightly smaller, while also more stable in-game now.