Rework internal depency counter logic to fix flaws (WIP)
Unlike the guid ref block which should contain all guid refs, the internal guids (guids of assets that reside in our pak file) must only have the guid counted once. Some assets like materials and animrigs can use an asset more than once, and if this asset happens to be in our pak, we should only count it once.

The implementation is not finished yet, there is some pending work to finalize this.