Implement new page and slab management class
New redesigned page and slab management class that will replace the existing one, this class fixes issues where pages and slabs aren't written to their full alignment. The control flow has also been improved and when we merge pages, we check on the full aligned page size (if requested alignment is higher than the page we check on, we use that alignment instead).