Pad the lump size to its aligned size
This was the last thing we always missed, just like how slabs and pages are padded out to their full aligned size, the lumps should be too. FindOrCreatePage and FindOrCreateSlab now expect to receive the full aligned size as we have to check on this when merging pages. This seems to fully fix the issue where decoded pointers in the runtime doesn't point to the correct data in pages.