Fix incorrect lump padding in page builder
The logic for padding the previous lump to align our current one was flawed as the align amount was calculated after the page size was increased. Slightly changed the code to have the find functions only return the slabs/pages without adding the sizes to them, and have CreatePageLump deal with that instead.

For the second issue; in commit 5c5cc00cc456bd14ac2732d1a29075bd3aeb7e80, a fix was implemented to pad the asset out to its alignment boundary. However, this was placed before the lump and not after. This padding lump was supposed to pad the data lump to its alignment boundary. It has been moved to after the data lump has been created.