Simplify pointer assignment and registration API
Make an API that registers and sets the pointer to the struct in a single call. The function only expects the objects now and deals with the calls internally. This significantly improves the readability of the code, and also significantly reduces the chances for human error where the previous GetPointer call for example as called on a different struct member than AddPointer. Now it is only defined once on call site.