Fix bug in slab data size calculation
This was always a bug in repak and never accounted for; the slabs sizes must also count in all padding required to pad a page to its alignment boundary, and padding required to pad a previous page with a lower alignment to align the page that comes next to it with a higher alignment. This fixes all corruption bugs as not adhering to this will result in the runtime allocating a slab buffer that is smaller than required to fit all pages in with aforementioned alignment rules.

Another change: pages inside pak files are not padded out to their full alignment boundaries, nor are the sizes. For slabs, this is identical. The runtime does this so we can save on some bytes here.