Improve texture classification code and support all textures of the game
The old code did not support 16 bit textures as we only checked on 8 and 32. Some textures of the game are of type r8g8, and there are probably a few others too. The classification code has been replaced with what texdiag uses in the Microsoft's DirectXTex library as it is very robust and should make sure we don't miss anything. If the user provides an unsupported texture, the tool will still throw out an error for this.

Also did some light optimizations to Texture_InternalAddTexture where we only lookup the texture format once.