BinaryIO class file size calculation bug fix
When seeking back and writing data, the size was still incremented while the actual file size never increased. If we seeked backwards, we have to keep track of this and only update once we are past the end of the stream.

The reason we do this instead of using tellp is because this is way cheaper and avoids an internal flush call.