Make padding consistent on all operating systems
Seeking beyond the end of the stream doesn't necesarily pad it. Some versions of operating systems or platforms might yield different behavior. On my system, this occasionally failed, especially when the positive seek offset from the end of the stream was large. Changed any padding behavior with an actual cheap Pad() call to make sure this will never be an issue.