Fix regression in material adding code
When we call MaterialAsset_t ::FromJSON, the asset version must be initialized because only version >= 15 has support for uberBufferFlags, Titanfall 2 materials (v12) does not. It is a required field. Since the version was 0 it would never set it for apex materials, and this broke particle materials as they need to set a certain flag to select a different uber buffer which is dedicated to particle materials (PTCU/PTCS).