Improve material shader type selection performance and add error handling
Use a switch case for this instead of a map, and check if the string is at least 3 since otherwise we would overflow it when reading 4 bytes for the int to select the type internally. Added an error if user happened to provide a shader type that is not supported as otherwise it would crash in the game.