Properly handle exceptions when a cell couldn't be parsed from the datatable
Instead of a silent failure (before this code was refactored to use a stream directly), or a crash (current), return a useful error msg indicating which cell couldn't be parsed. This should fully eliminate any possiblities for errors in a data table CSV file which may affect both the tool or the runtime.

Also improved existing error messages to print the cell in the order of: row, column as this makes more sense.