Fully implement starpak data deduplication
The systems added in commit 396b044b4e1f28f96f3b0667bf9c0a0c974b46bd and v have now been fully implemented.

The system works by parsing a precomputed map file containins the data hashes, starpak names and offsets into them. It hashes the asset we are trying to add to the starpak and compares if it already exists. If it exists we just map the offsets and indices to that starpak order to deduplicate it.

The system also keeps track of newly added assets so that when the same asset is added again, it would also get deduplicated.

This patch required a large change to the code base as the starpak creation code was inline with the pak build class. It needs to be separate and persistent accross individual rpak builds.

The tool now supports building a list of paks as well, which is where the deduplication works the best with. Individual paks can still be built using the classic way.