Log the asset name with the warning or error
Since debug logs are disabled by default, and almost all errors only print the actual without the asset name, we must add it in.

The reason errors never printed the asset name along was because this was previously logged when the asset was to be added, so one could look up what the last added asset was to find the origin of the error.

However, even the 'Adding <asset>' print slowed the application down by a considerable amount so this was also moved under the debug option.