Implement ability to override streaming layout and resource flags
User can now determine what needs to be permanent, mandatory or optional. This is extremely powerful as it allows us to deduplicate streamed data even more. This dropped a test case from 6gb down to 1.8gb. Resource flags are also needed as cubemaps have different flags and that needs to be handled correctly.