DataTable asset improvements
Implement support for guid refs on precached assets, this allows the game to resolve all the dependencies first (and precache them automatically) instead of having it being done manually by the user. Also made sure that the value buffer is always aligned to the containing data types with the highest alignment.