Improve material errors and warnings
The previous errors were confusing, as it was unclear what the texture index and bind point was, this has now been clarified. The warning was confusing as well, as it could be interpretted as an error while in fact if a texture wasn't present on the disk at the time of adding this material, repak will treat it as an external reference which is expected behavior. Reworked the warning to make this clear.