Fix deduplication failures caused by unaligned data buffers
The deduplication code must receive a buffer aligned and padded out to the size of STARPAK_DATABLOCK_ALIGNMENT, since the data within all other starpaks are aligned and padded out to this value, and the hashing algorithm hashes the entire thing. Leaving the padding out will therefore yield different results and cause data to be not deduplicated. Fixed by padding everything out and added a dedicated warning may it ever receive something unaligned in the future.