NULL g_currentAsset if we ended up adding the asset
This wouldn't get nulled if the asset was added, and if we error outside the asset packing handler, it would show the last asset even though we are done with that asset and the current error has nothing to do with that asset. Properly null it to make sure this doesn't happen.