Improve build list UX
If we couldn't find the build list, or if there was a parsing error, the current build list file wouldn't be printed. JSON_ParseFromFile() was made to print the asset from g_currentAsset , however since its used here there won't be a current asset since this is the pak itself. So print out the current build map we are going to parse to allow figuring out which build map in the build list is bad.