Properly classify and register model and settings dependencies
Model and settings dependencies need to be explicitly defined; a pointer needs to be set to the start of the model dependencies GUID, along with its count. Settings assets have the exact same process. The runtime uses this to precache models and all other assets inside the settings assets when an animation is getting loaded and played. This patch allows common.rpak to be fully rebuilt while maintaining original quality and behavior.