Remove extraneous duplication check
Check for duplicate GUID's (collisions or user errors) already takes place within BeginAsset(), doing it in AddJSONAsset() is extraneous, and also will not guarantee full coverage as the auto-add feature bypasses this check all together. The check in BeginAsset() guarantees full coverage. Also did some improvements to the existing error to report whether we have a collision or when the user has made a mistake.