Match LCD field names with shader structure ( CBufEnginePost )
Match the RLCD asset type field names with the member field names found in the CBufEnginePost shader class (this is indirectly mapped to the shader). This allows for clearer documentation and avoids confusion when looking up the CBufEnginePost member variables and comparing them with the ones found in LcdScreenEffect_s.