Major improvements to BinaryIO
BinaryIO now returns true if read and write operations were successful, false if they weren't. The size housekeeping logic has also been rewritten. If we seek forward, we will store this amount now and only add it to the total size if we end up writing something at that location. Previously it would always add to the totalSize even if we weren't writing to the seeked pos which is incorrect!